Add files via upload
This is my first python project I have done . Is a quiz game specifically a python quiz game .  It asks various questions related to python and state weather you are write or wrong by the feedback it gives. I made used of if  and else statements, lower(), upper(), quit() and also declared a variable . I also concatenated an integer and a string and so on. So basically this is what my first python project is all about.